Exercícios de 005 a 013 corrigidos. Exercícios 014 e 015 feitos e corrigidos

+++ Mundo01/Exercicios/ex007.py
@@ -1,5 +1,5 @@
-n1 = int(input("Coloque sua primeira nota: "))
-n2 = int(input("Coloque sua segunda nota: "))
+n1 = float(input("Coloque sua primeira nota: "))
+n2 = float(input("Coloque sua segunda nota: "))
 
 media = (n1 + n2)/2
